+++ .eslintrc.js
@@ -66,6 +66,7 @@ module.exports = {
         tsx: 'never',
         js: 'never',
         jsx: 'never',
+        json: 'always',
       },
     ],
     'import/no-cycle': 'off',