+++ docs/docs/sandpack/code-deps.ts
@@ -1,5 +1,24 @@
+import {
+  excalidrawVersion,
+  juiceVersion,
+  plateVersion,
+  testUtilsVersion,
+} from './plate-versions';
+
+export const plateTestUtilsDeps = {
+  '@udecode/plate-test-utils': testUtilsVersion,
+};
+
+export const excalidrawDeps = {
+  '@udecode/plate-ui-excalidraw': excalidrawVersion,
+};
+
+export const juiceDeps = {
+  '@udecode/plate-juice': juiceVersion,
+};
+
 export const commonDeps = {
-  '@udecode/plate': '*',
+  '@udecode/plate': plateVersion,
   react: '17.0.2',
   'react-dom': '17.0.2',
   'react-scripts': '4.0.3',
@@ -12,28 +31,11 @@ export const commonDeps = {
   'react-dnd-html5-backend': '15.1.3',
 };
 
-export const plateTestUtilsDeps = {
-  '@udecode/plate-test-utils': '*',
-};
-
 export const toolbarDeps = {
   '@styled-icons/boxicons-regular': '10.23.0',
   '@styled-icons/foundation': '10.28.0',
   '@styled-icons/material': '10.28.0',
   '@tippyjs/react': '4.2.5',
-};
-
-export const excalidrawDeps = {
-  '@udecode/plate-ui-excalidraw': '*',
-};
-
-export const juiceDeps = {
-  '@udecode/plate-juice': '*',
-};
-
-export const zustandDeps = {
-  '@udecode/zustood': '1.1.1',
-  zustand: '^3.7.2',
 };
 
 export const playgroundDeps = {
@@ -42,5 +44,4 @@ export const playgroundDeps = {
   ...toolbarDeps,
   ...excalidrawDeps,
   ...juiceDeps,
-  ...zustandDeps,
 };