+++ docs/docs/sandpack/files/code-PlaygroundApp.tsx
@@ -1,5 +1,4 @@
-export const playgroundAppCode = `import 'tippy.js/dist/tippy.css';
-import React, { CSSProperties, useMemo, useRef } from 'react';
+export const playgroundAppCode = `import React, { CSSProperties, useMemo, useRef } from 'react';
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
 import {
@@ -103,9 +102,7 @@ const App = () => {
           createBlockquotePlugin(),
           createTodoListPlugin(),
           createHeadingPlugin(),
-          createImagePlugin({
-            type: 'image-block',
-          }),
+          createImagePlugin(),
           createHorizontalRulePlugin(),
           createLinkPlugin(),
           createListPlugin(),