+++ docs/docs/sandpack/files/common/code-editableProps.ts
@@ -4,6 +4,7 @@ import { MyValue } from '../typescript/plateTypes';
 export const editableProps: TEditableProps<MyValue> = {
   spellCheck: false,
   autoFocus: false,
+  readOnly: false,
   placeholder: 'Type…',
 };
 `;