+++ packages/core/src/utils/misc/index.ts
@@ -13,3 +13,4 @@ export * from './hotkeys';
 export * from './isUrl';
 export * from './jotai';
 export * from './type-utils';
+export * from './zustood';