+++ packages/nodes/image/src/createImagePlugin.ts
@@ -1,4 +1,10 @@
-import { createPluginFactory } from '@udecode/plate-core';
+import {
+  createPluginFactory,
+  getBlockAbove,
+  getPluginType,
+} from '@udecode/plate-core';
+import isHotkey from 'is-hotkey';
+import { imageGlobalStore } from './components/index';
 import { ImagePlugin } from './types';
 import { withImageUpload } from './withImageUpload';
 
@@ -12,6 +18,23 @@ export const createImagePlugin = createPluginFactory<ImagePlugin>({
   isElement: true,
   isVoid: true,
   withOverrides: withImageUpload,
+  handlers: {
+    onKeyDown: (editor) => (e) => {
+      // focus caption from image
+      if (isHotkey('down', e)) {
+        const entry = getBlockAbove(editor, {
+          match: { type: getPluginType(editor, ELEMENT_IMAGE) },
+        });
+        if (!entry) return;
+
+        imageGlobalStore.set.focusEndCaptionPath(entry[1]);
+      }
+
+      // TODO: focus caption from line below image
+      // if (isHotkey('up', e)) {
+      // }
+    },
+  },
   then: (editor, { type }) => ({
     deserializeHtml: {
       rules: [