+++ packages/nodes/list/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @udecode/plate-list
 
+## 13.8.0
+
+### Patch Changes
+
+- [#1651](https://github.com/udecode/plate/pull/1651) by [@davisg123](https://github.com/davisg123) – Some types of nested lists are not unwrapped completely
+
+## 13.7.0
+
+## 13.6.0
+
+### Patch Changes
+
+- [`42c5ed1`](https://github.com/udecode/plate/commit/42c5ed1a15dfecb7e64fc39ba328c16733472112) by [@zbeyens](https://github.com/zbeyens) – Toggling a nested list that includes paragraph content should toggle all nested list elements
+
 ## 13.5.0
 
 ## 13.3.1